+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+Fixes:
+- Fix Belgium to ignore `-` and `.` 🇧🇪
+
 ## [1.3.0] - 2025-12-12
 
 Features: